+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Unreleased - YYYY-MM-DD
 - Added io/file package
 - Added archive/zipfile package
+- Added archive/tgzfile package